requirements.txt
anthropic>=0.40.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
diffusers>=0.30.0
transformers>=4.44.0
accelerate>=0.33.0
sentencepiece
protobuf
opencv-python-headless
Pillow>=10.0.0
imagehash>=4.3.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
kokoro>=0.9.4
soundfile
numpy
huggingface-hub
ftfy
